+++ src/test/java/com/example/i_commerce/domain/member/service/SellerServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.i_commerce.domain.member;
+package com.example.i_commerce.domain.member.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
